requirements.txt
python-telegram-bot[job-queue]==20.8
SQLAlchemy==2.0.36
requests==2.32.3
pytz==2024.2
matplotlib==3.9.2
python-dotenv==1.0.1
pytest==8.3.3
